Cleanup and clarify test
